+++ example/android/app/src/main/kotlin/com/example/advanced_image_processing_toolkit_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.advanced_image_processing_toolkit_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()